+++ testing/image_testing.py
@@ -1,1 +1,0 @@
-from ATK.structures.data_containers.definitions import Image