+++ changelog.md
@@ -9,3 +9,4 @@ Changes
 To-Do 
 -----
 - Let Vizier catalogue names be used in overlays (as a fallback if alias not in alias file)
+- use astropy units to define axes